requirements.txt
fastapi
uvicorn
sqlalchemy
aiosqlite
passlib[bcrypt]
python-jose[cryptography]
email-validator
pydantic[email]
authlib
httpx